Change the namespace used by the package
Old: com.reactnativedocumentpicker
New: io.github.elyx0.reactnativedocumentpicker
