+++ android/src/main/AndroidManifest.xml
@@ -1,6 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-      package="com.reactnativedocumentpicker"
-      android:versionCode="1"
-      android:versionName="1.0">
+      package="io.github.elyx0.reactnativedocumentpicker">
 </manifest>
