--- android/src/main/java/com/reactnativedocumentpicker/ReactNativeDocumentPicker.java
@@ -1,32 +1,0 @@
-package com.reactnativedocumentpicker;
-
-import com.facebook.react.ReactPackage;
-import com.facebook.react.bridge.JavaScriptModule;
-import com.facebook.react.bridge.NativeModule;
-import com.facebook.react.bridge.ReactApplicationContext;
-import com.facebook.react.uimanager.ViewManager;
-import com.reactnativedocumentpicker.DocumentPicker;
-
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-
-public class ReactNativeDocumentPicker implements ReactPackage {
-
-    // Deprecated RN 0.47
-    public List<Class<? extends JavaScriptModule>> createJSModules() {
-        return Collections.emptyList();
-    }
-
-    @Override
-    public List<ViewManager> createViewManagers(ReactApplicationContext reactContext) {
-        return Collections.emptyList();
-    }
-
-    @Override
-    public List<NativeModule> createNativeModules(ReactApplicationContext reactContext) {
-        List<NativeModule> modules = new ArrayList<>();
-        modules.add(new DocumentPicker(reactContext));
-        return modules;
-    }
-}
