+++ android/src/main/java/io/github/elyx0/reactnativedocumentpicker/ReactNativeDocumentPicker.java
@@ -1,0 +1,31 @@
+package io.github.elyx0.reactnativedocumentpicker;
+
+import com.facebook.react.ReactPackage;
+import com.facebook.react.bridge.JavaScriptModule;
+import com.facebook.react.bridge.NativeModule;
+import com.facebook.react.bridge.ReactApplicationContext;
+import com.facebook.react.uimanager.ViewManager;
+
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+
+public class ReactNativeDocumentPicker implements ReactPackage {
+
+	// Deprecated RN 0.47
+	public List<Class<? extends JavaScriptModule>> createJSModules() {
+		return Collections.emptyList();
+	}
+
+	@Override
+	public List<ViewManager> createViewManagers(ReactApplicationContext reactContext) {
+		return Collections.emptyList();
+	}
+
+	@Override
+	public List<NativeModule> createNativeModules(ReactApplicationContext reactContext) {
+		List<NativeModule> modules = new ArrayList<>();
+		modules.add(new DocumentPicker(reactContext));
+		return modules;
+	}
+}
